Split up the rust lib into modules
	This is done as preparation for coming changes to this module.
	The future changes in this branch will include:
		- reading of the signal header
		- reading of the data
		- the Reader will return the full edf file stucture
		  instead of the header
