+++ src/lib.rs
@@ -1,5 +1,6 @@
-pub use crate::error::{Error, ErrorKind};
-pub use crate::reader::Reader;
-
 mod error;
 mod reader;
+mod edf_file;
+
+pub use crate::error::{Error, ErrorKind};
+pub use crate::reader::Reader;
